pyproject.toml
[project]
authors = [{name = "Matthias Homann", email = "palto@mailbox.org"}]
classifiers = [
  "Intended Audience :: Developers",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
  "loguru>=0.7.3",
  "psutil>=7.0.0",
  "pyyaml>=6.0.2",
  "requests>=2.32.3",
]
description = "This tool downloads and installs release binaries from GitHub repositories."
keywords = ['python']
name = "girsh"
readme = "README.md"
requires-python = ">=3.10,<4.0"
version = "0.1.30"

[project.urls]
Documentation = "https://palto42.github.io/girsh/"
Homepage = "https://palto42.github.io/girsh/"
Repository = "https://github.com/palto42/girsh"

[dependency-groups]
dev = [
  "pytest>=7.2.0",
  "pre-commit>=2.20.0",
  "tox-uv>=1.11.3",
  "deptry>=0.23.0",
  "mypy>=0.991",
  "pytest-cov>=4.0.0",
  "ruff>=0.11.5",
  "mkdocs>=1.4.2",
  "mkdocs-material>=8.5.10",
  "mkdocstrings[python]>=0.26.1",
  "types-pyyaml>=6.0.12.20241230",
  "types-requests>=2.32.0.20241016",
  "types-psutil>=6.1.0.20241221",
]

[tool.setuptools.package-data]
"girsh.templates" = ["config_template.yaml"]

[project.scripts]
girsh = "girsh.girsh:main"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
files = ["src"]
no_implicit_optional = true
show_error_codes = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
fix = true
line-length = 120
target-version = "py311"

[tool.ruff.lint]
ignore = [
  # LineTooLong
  "E501", # DoNotAssignLambda
  "E731",
]
select = [
  # flake8-2020
  "YTT", # flake8-bandit
  "S", # flake8-bugbear
  "B", # flake8-builtins
  "A", # flake8-comprehensions
  "C4", # flake8-debugger
  "T10", # flake8-simplify
  "SIM", # isort
  "I", # mccabe
  "C90", # pycodestyle
  "E",
  "W", # pyflakes
  "F", # pygrep-hooks
  "PGH", # pyupgrade
  "UP", # ruff
  "RUF", # tryceratops
  "TRY",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["src"]
